dev/requirements.txt
PyYAML==6.0
numpy==1.24.0
tensorflow==2.14.0
tensorflow-estimator==2.14.0
requests[security]
